Moved server templates into own file
