+++ cmd/protoc-gen-gorums/gengorums/gorums.go
@@ -205,9 +205,10 @@ func callTypeName(ext *protoimpl.ExtensionInfo) string {
 // The entries in this map is used to generate dev/zorums_{type}.pb.go
 // files for the different keys.
 var gorumsCallTypesInfo = map[string]*callTypeInfo{
-	"node":  {template: node},
-	"qspec": {template: qspecInterface},
-	"types": {template: datatypes},
+	"node":   {template: node},
+	"qspec":  {template: qspecInterface},
+	"types":  {template: datatypes},
+	"server": {template: server},
 
 	callTypeName(gorums.E_Quorumcall): {
 		extInfo:  gorums.E_Quorumcall,
